Update promptbase.mmlu (#23)
Working to get `promptbase.mmlu` into a runnable form. Changes include:

- Adjust `format_mmlu.py` script so that it has command line arguments
- Add `mmlu` as an allowed dataset. This permits commands like `python -m promptbase mmlu --subject clinical_knowledge`
- Try to impose some regularity on where intermediate files are stored

Several of the files have also been reformatted as part of this, which does make the changes look slightly more extensive than they really are.

So far, only the `generate()` side of things has been worked on (and is not currently fully working). The evaluation is still to come.