AzureML pipeline with kNN fewshot CoT (#40)
Add an AzureML pipeline which provides some examples with chain-of-thought in the main prompt. Starting with an 'example' dataset and a 'test' dataset:

1. Generate CoTs and answers for the example dataset with a zero-shot prompt
2. Filter out examples which had incorrect answers
3. Compute embedding vectors for the 'test' dataset and remaining examples
4. Use cosine similarity to select kNN examples for each entry in the 'test' dataset
5. Use the kNN examples in a few-shot CoT prompt for the test dataset

The `azureml_pipelines.py` file is also getting unwieldly. There is probably a template-pattern refactoring lurking in there, but I have not yet figured out how to do that.